# ExtractImagePatches gradient refactoring (XLA + speedup)