Fixes https://github.com/tensorflow/tensorflow/issues/59058 and https://github.com/tensorflow/tensorflow/issues/59061
Improves https://github.com/tensorflow/tensorflow/issues/20146

Colab + CPU:

| Implementation | CPU user |  CPU sys | CPU total | Wall time |
|:---------------|---------:|---------:|----------:|----------:|
| Current        |   2.80 s |  1.73  **s** |    4.54 s |    5.39 s |
| Proposed       |   2.87 s | 42.20 **ms** |    2.91 s |    3.05 s |

Colab + GPU

| Implementation | CPU user |  CPU sys | CPU total | Wall time |
|:---------------|---------:|---------:|----------:|----------:|
| Current        | 907.0 ms | 266.0 ms |   1.17  **s** |    830 ms |
| Proposed       |  17.4 ms |   2.1 ms |  19.50 **ms** |   56.7 ms |